+++ CHANGELOG.md
@@ -6,15 +6,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.3] - 2017-06-20
+## [2.0.0] - 2017-06-26
+### Added
+- Added support for model parameters of `make:controller`, `make:factory`, `make:observer` and `make:policy`
+
+### Removed
+- Laravel 5.8 support
+
+### Changed
+- Replace extended class with macro and afterResolving hooks
+
+## [1.0.3] - 2017-06-24
 ### Removed
 - Production mode checking to make comply with the Laravel way.
 
-## [1.0.2] - 2017-06-20
+## [1.0.2] - 2017-06-24
 ### Added
 - Support for Laravel 5.8
 
-## [1.0.1] - 2017-06-20
+## [1.0.1] - 2017-06-24
 ### Added
 - This changlog
 - MIT license
@@ -22,6 +32,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - The Provider does not return the model:make command any more in production environment
 
-## [1.0.0] - 2017-06-20
+## [1.0.0] - 2017-06-24
 ### Added
 - Initial version implementing the configurable model make command